% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMaterial.R
\name{addMaterial}
\alias{addMaterial}
\title{Adding new material data to material database.}
\usage{
addMaterial(name, reference, materialIsotropic, youngModulus = NA,
  poissonRatio = NA, density, flexuralRigidity11 = NA,
  flexuralRigidity22 = NA, flexuralRigidity12 = NA,
  flexuralRigidity66 = NA, transverseShearmodulus44 = NA,
  transverseShearmodulus55 = NA)
}
\arguments{
\item{name}{The material name.}

\item{reference}{The material data reference.}

\item{materialIsotropic}{Logical, \code{TRUE} means material is isotropic.}

\item{youngModulus}{The material Young modulus in Pa, Default: NA}

\item{poissonRatio}{The material Poisson ratio, Default: NA}

\item{density}{The material density in kg/m3}

\item{flexuralRigidity11}{The material flexural rigidity 11 in Pa m3, Default: NA}

\item{flexuralRigidity22}{The material flexural rigidity 22 in Pa m3, Default: NA}

\item{flexuralRigidity12}{The material flexural rigidity 12 in Pa m3, Default: NA}

\item{flexuralRigidity66}{The material flexural rigidity 66 in Pa m3, Default: NA}

\item{transverseShearmodulus44}{The material transverse shear modulus 44 in MPa m, Default: NA}

\item{transverseShearmodulus55}{The material transverse shear modulus 55 in MPa m, Default: NA}
}
\value{
Adds the new material data to the material database embedded into the \code{pongaR} library.
}
\description{
A utility function that adds new material data to \code{pongaR} material database.
}
