% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apertureFunction.R
\name{apertureFunction}
\alias{apertureFunction}
\title{Computing sensor's aperture function.}
\usage{
apertureFunction(x, gamma_m, alpha_n, x0, y0, receiverRadius2)
}
\arguments{
\item{x}{A distance from the center along the sensor's radius, in m.}

\item{gamma_m}{A parameter from Midlin's plate theory.}

\item{alpha_n}{A parameter from Midlin's plate theory.}

\item{x0}{Sensor position coordinate \code{x0}, in m.}

\item{y0}{Sensor position coordinate \code{y0}, in m.}

\item{receiverRadius2}{Squared sensor radius, in m2.}
}
\value{
The value of the sensor's aperture function at a distance \code{x} from the center.
}
\description{
A utility function that computes sensor's aperture function.
}
