% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_A.R
\name{delta_A}
\alias{delta_A}
\title{Lamb wave dispersion function for antisymmetric modes.}
\usage{
delta_A(x, par)
}
\arguments{
\item{x}{The wavenumber}

\item{par}{The parameters thickness, longitudinal velocity, transvers velocity and angular frequency
as a \code{list} object.}
}
\value{
The value of the Lamb wave dispersion function for antisymmetric modes according
        Wilcox's PhD thesis eq. 2.15, p. 28.
}
\description{
A utility function that calculates the Lamb wave dispersion function for antisymmetric modes according
             Wilcox's PhD thesis eq. 2.15, p. 28.
}
