% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_S.R
\name{delta_S}
\alias{delta_S}
\title{Lamb wave dispersion function for symmetric modes.}
\usage{
delta_S(x, par)
}
\arguments{
\item{x}{The wavenumber}

\item{par}{The parameters thickness, longitudinal velocity, transvers velocity and angular frequency
as a \code{list} object.}
}
\value{
The value of the Lamb wave dispersion function for symmetric modes according
        Wilcox's PhD thesis eq. 2.15, p. 28.
}
\description{
A utility function that calculates the Lamb wave dispersion function for symmetric modes
             according Wilcox's PhD thesis eq. 2.15, p. 28.
}
