% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Extracting acoustic emission features.}
\usage{
getFeatures(wave = makeSource()$waveform, xlim = c(0, duration(wave)),
  samp.rate = 1e+07, preAmpGain = 40, preTriggerSamples = 0,
  threshold = 0.01, maxDuration = 0.05, maxHitNum = 5000,
  HDT = 1e-04, HLT = 1e-05, PDT = 5e-05, powerLimits = rbind(c(0,
  250000), c(250000, 5e+05), c(5e+05, 750000), c(750000, 1e+06)),
  plot = FALSE)
}
\arguments{
\item{wave}{The waveform as a \code{Wave} object, Default: makeSource()$waveform}

\item{xlim}{The x limits (x1, x2) of the plot. Note that x1 > x2 is allowed and leads
to a ‘reversed axis’., Default: c(0, duration(wave))}

\item{samp.rate}{Sampling rate of the \code{Wave} in Hz, Default: 1e+07}

\item{preAmpGain}{Pre-amplifier gain in dB, Default: 40}

\item{preTriggerSamples}{Number of pre-trigger samples, Default: 0}

\item{threshold}{Thershold in V, Default: 0.01}

\item{maxDuration}{Max hit duration in s, Default: 0.05}

\item{maxHitNum}{Maximum number of hits extracted, Default: 5000}

\item{HDT}{Hit definition time in s, Default: 1e-04}

\item{HLT}{Hit lockout time in s, Default: 1e-05}

\item{PDT}{Peak definition time in s, Default: 5e-05}

\item{powerLimits}{Ranges of the four partial power features,
Default: rbind(c(0, 250000), c(250000, 5e+05), c(5e+05, 750000), c(750000, 1e+06))}

\item{plot}{If \code{TRUE}, plots the waveform with indications of the threshold
and some features, Default: FALSE}
}
\value{
A \code{list} object with the extracted acoustic emission features for each
        detected hit in the waveform. It also includes the start, peak and end indexes of the hits.
}
\description{
Extracts several acoustic emission features from a waveform as defined in
             Sause's PhD thesis table 2.2, p. 50.
}
