% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lame2vel.R
\name{lame2vel}
\alias{lame2vel}
\title{Longitudinal and transverse velocities from Lame parameters.}
\usage{
lame2vel(density = 4460, lameLambda = 70725788000,
  lameMu = 46530734000)
}
\arguments{
\item{density}{The material density in kg/m3, Default: 4460}

\item{lameLambda}{The material Lame first parameter in Pa, Default: 70725788000}

\item{lameMu}{The material Lame second parameter in Pa, Default: 46530734000}
}
\value{
The material density in kg/m3, longitudinal and transverse velocities in m/s as
        a \code{list} object.
}
\description{
A utility function that converts Lame parameters to longitudinal and transverse velocities.
}
