% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lame2young.R
\name{lame2young}
\alias{lame2young}
\title{Young modulus and Poisson ratio from Lame parameters.}
\usage{
lame2young(lameLambda = 70725788000, lameMu = 46530734000)
}
\arguments{
\item{lameLambda}{The material Lame first parameter in Pa, Default: 70725788000}

\item{lameMu}{The material Lame second parameter in Pa, Default: 46530734000}
}
\value{
The material Young modulus in Pa and the Poisson ratio as
        a \code{list} object.
}
\description{
A utility function that converts Lame parameters to Young modulus and Poisson ratio.
}
