% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDD.R
\name{makeDD}
\alias{makeDD}
\title{Generating dispersion data.}
\usage{
makeDD(phaseData.file = "titanium.txt", thickness = 0.001,
  density = 4460, longVelocity = 6060, tranVelocity = 3230,
  plot = FALSE)
}
\arguments{
\item{phaseData.file}{A text file with exported dispersion data from Disperse , Default: "titanium.txt"}

\item{thickness}{The plate thickness in m, Default: 0.001}

\item{density}{The material density in kg/m3, Default: 4460}

\item{longVelocity}{The material longitudinal velocity in m/s, Default: 6060}

\item{tranVelocity}{The material transverse velocity in m/s, Default: 3230}

\item{plot}{If \code{TRUE}, plots the dispersion data, Default: FALSE}
}
\value{
A dataset with dispersion data embedded into the \code{pongaR} library.
}
\description{
A function that generates dispersion data.
}
