% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEF.R
\name{makeEF}
\alias{makeEF}
\title{Generating excitability function data.}
\usage{
makeEF(dispersionData = "titaniumDD", plot = FALSE)
}
\arguments{
\item{dispersionData}{The dispersion data dataset, Default: 'titaniumDD'}

\item{plot}{If \code{TRUE}, plots the excitability function data, Default: FALSE}
}
\value{
A dataset with excitability function data embedded into the \code{pongaR} library.
}
\description{
A function that generates excitability function data.
}
\seealso{
\code{\link[numDeriv]{grad}}
}
