% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFR.R
\name{makeFR}
\alias{makeFR}
\title{Generating sensor frequency response data.}
\usage{
makeFR(sensorCalib = "r15iuc.csv", xunit = c("Hz", "kHz", "MHz"),
  yunit = c("V/ubar", "V/(m/s)", "V/m"), dBref = 1, n = 4097,
  plot = FALSE)
}
\arguments{
\item{sensorCalib}{The .csv file with the digitized  calibration  curve, Default: 'r15iuc.csv'}

\item{xunit}{Character indicating which unit is used in for the frequency axis, Default: c("Hz", "kHz", "MHz")}

\item{yunit}{Character indicating which unit is used in for the magnitude axis, Default: c("V/ubar", "V/(m/s)", "V/m")}

\item{dBref}{A dB reference value, Default: 1}

\item{n}{Number of frequency response points, Default: 4097}

\item{plot}{If \code{TRUE}, plots the sensor frequency response data, Default: FALSE}
}
\value{
A dataset with sensor frequency response data embedded into the \code{pongaR} library.
}
\description{
A function that generates sensor frequency response data.
}
