% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMedium.R
\name{makeMedium}
\alias{makeMedium}
\title{Generating medium impulse response.}
\usage{
makeMedium(plate = makePlate(), plateTheory = c("lamb", "mindlin",
  "kirchhoff"), excitabilityData = "titaniumEF", modes = NULL,
  sensorDiameter = 0, sensorPosition = c(0.254, 0.19),
  sourcePosition = c(0.204, 0.19), sourceType = c("impulse", "step"),
  duration = 0.0001024, xunit = c("time", "samples"),
  samp.rate = 1e+07, bit = 32)
}
\arguments{
\item{plate}{Plate specification, Default: makePlate()}

\item{plateTheory}{Plate theory model, Default: c("lamb", "mindlin", "kirchhoff")}

\item{excitabilityData}{Excitability function data, Default: 'titaniumEF'}

\item{modes}{Maximum number of modes \code{c(m,n)}, Default: NULL}

\item{sensorDiameter}{Sensor diameter (in m), Default: 0}

\item{sensorPosition}{Sensor position coordinates \code{c(x,y)} in m, Default: c(0.254, 0.19)}

\item{sourcePosition}{Sensor position coordinates \code{c(x0,y0)} in m, Default: c(0.204, 0.19)}

\item{sourceType}{Source type, Default: c("impulse", "step")}

\item{duration}{Duration of the \code{Wave} in \code{xunit}, Default: 0.0001024}

\item{xunit}{Character indicating which unit is used in argument \code{duration}.
If \code{xunit = "time"}, the unit is time in s,
otherwise the number of samples, Default: c("time", "samples")}

\item{samp.rate}{Sampling rate of the \code{Wave}, in Hz, Default: 1e+07}

\item{bit}{Resolution of the \code{Wave} and rescaling unit, Default: 32. This may be
1 for rescaling to numeric values in [-1,1],
8 (i.e. 8-bit) for rescaling to integers in [0, 254],
16 (i.e. 16-bit) for rescaling to integers in [-32767, 32767],
24 (i.e. 24-bit) for rescaling to integers in [-8388607, 8388607],
32 (i.e. 32-bit) for rescaling either to integers in [-2147483647, 2147483647],
64 (i.e. 64-bit) for rescaling to numeric values in [-1, 1] (FLOAT_IEEE Wave format), and
0 for not rescaling at all.}
}
\value{
The medium impulse response waveform as a \code{medium} object.
}
\description{
A function that generates medium impulse response.
}
