% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReadout.R
\name{makeReadout}
\alias{makeReadout}
\title{Generating readout impulse response.}
\usage{
makeReadout(sensorName = "r15iuc", filterRange = c(NULL, NULL),
  amplifierGain = 0, plot = FALSE)
}
\arguments{
\item{sensorName}{The sensor dataset name, Default: 'r15iuc'}

\item{filterRange}{The band-pass filter range, in Hz, Default: c(NULL, NULL)}

\item{amplifierGain}{The amplfier gain in dB, Default: 0}

\item{plot}{If \code{TRUE}, plots the readout impulse response, Default: FALSE}
}
\value{
The readout impulse response waveform as a \code{readout} object.
}
\description{
A function that generates readout impulse response.
}
