% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRun.R
\name{makeRun}
\alias{makeRun}
\title{Generating run data.}
\usage{
makeRun(expData = makeExp())
}
\arguments{
\item{expData}{An object of class \code{exp} produced by \code{makeExp}, Default: makeExp()}
}
\value{
The run data as a \code{run} object.
}
\description{
A function that generates run data to produce several acoustic emission signal
             waveforms defined by \code{expData}.
}
