% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignal.R
\name{makeSignal}
\alias{makeSignal}
\title{Generating acoustic emission signal due solid particle impact onto a plate.}
\usage{
makeSignal(runData = NULL, sourceData = NULL, mediumData = NULL,
  readoutData = NULL, plot = FALSE)
}
\arguments{
\item{runData}{An object of class \code{run} produced by \code{makeRun}, Default: NULL}

\item{sourceData}{An object of class \code{source} produced by \code{makeSource}, Default: NULL}

\item{mediumData}{An object of class \code{medium} produced by \code{makeMedium}, Default: NULL}

\item{readoutData}{An object of class \code{readout} produced by \code{makeReadout}}

\item{plot}{If \code{TRUE}, plots the readout impulse response, Default: FALSE}
}
\value{
The acoustic emission signal waveform as a \code{signal} object.
}
\description{
A function that generates acoustic emission signal due solid particle impact onto a plate.
}
