% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSource.R
\name{makeSource}
\alias{makeSource}
\title{Generating source impulse response.}
\usage{
makeSource(sourceModel = c("toneburst", "plb", "hertz", "hunter", "reed",
  "akay"), plate = NA, sphere = NA, riseTime = 1.5e-06,
  forceAmplitude = 1, centerFreq = 2e+05, cyclesNum = 5,
  duration = 0.0001024, xunit = c("time", "samples"),
  samp.rate = 1e+07, bit = 32)
}
\arguments{
\item{sourceModel}{The source model, Default: c("toneburst", "plb", "hertz", "hunter", "reed", "akay")}

\item{plate}{The plate specification produced by \code{makePlate}, Default: NA}

\item{sphere}{The solid sphere specification produced by \code{makeSphere}, Default: NA}

\item{riseTime}{The plb rise time, in s, Default: 1.5e-06}

\item{forceAmplitude}{The plb and tone burst force amplitude, in N, Default: 1}

\item{centerFreq}{The tone burst center frequency, in Hz, Default: 2e+05}

\item{cyclesNum}{The tone burst number of cycles, Default: 5}

\item{duration}{Duration of the \code{Wave} in \code{xunit}, Default: 0.0001024}

\item{xunit}{Character indicating which unit is used in argument \code{duration}.
If \code{xunit = "time"}, the unit is time in s,
otherwise the number of samples, Default: c("time", "samples")}

\item{samp.rate}{Sampling rate of the \code{Wave}, in Hz, Default: 1e+07}

\item{bit}{Resolution of the \code{Wave} and rescaling unit, Default: 32. This may be
1 for rescaling to numeric values in [-1,1],
8 (i.e. 8-bit) for rescaling to integers in [0, 254],
16 (i.e. 16-bit) for rescaling to integers in [-32767, 32767],
24 (i.e. 24-bit) for rescaling to integers in [-8388607, 8388607],
32 (i.e. 32-bit) for rescaling either to integers in [-2147483647, 2147483647],
64 (i.e. 64-bit) for rescaling to numeric values in [-1, 1] (FLOAT_IEEE Wave format), and
0 for not rescaling at all.}
}
\value{
The source impulse response waveform as a \code{source} object.
}
\description{
A function that generates source impulse response.
}
