% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSphere.R
\name{makeSphere}
\alias{makeSphere}
\title{Generating solid sphere specification.}
\usage{
makeSphere(diameter = 0.001, impactVelocity = 1,
  youngModulus = 74643767730, poissonRatio = 0.2285018,
  density = 2480)
}
\arguments{
\item{diameter}{PARAM_DESCRIPTION, Default: 0.001}

\item{impactVelocity}{PARAM_DESCRIPTION, Default: 1}

\item{youngModulus}{The material Young modulus in Pa, Default: 74643767730}

\item{poissonRatio}{The material Poisson ratio, Default: 0.2285018}

\item{density}{The material density in kg/m3, Default: 2480}
}
\value{
The plate specification as a \code{sphere} object.
}
\description{
A function that generates solid sphere specification.
}
