% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\docType{data}
\name{material}
\alias{material}
\title{\code{pongaR} material database}
\format{Material database as an object of class \code{"data.frame"}.}
\usage{
data(material)
}
\description{
Data from materials used in the simulations. To add new data, use function \code{addMaterial}
}
\keyword{datasets}
