% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r15i-uc.R
\docType{data}
\name{r15iuc}
\alias{r15iuc}
\title{Acoustic emission sensor frequency response}
\format{Frequency response as an object of class \code{"data.frame"}.}
\usage{
data(r15iuc)
}
\description{
Frequency response of a popular acoustic emission sensor.
}
\keyword{datasets}
