% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExp.R
\name{runExp}
\alias{runExp}
\title{Running experiment to produce acoustic emission signals due solid particle impacts onto a plate.}
\usage{
runExp(expName = "exp1", expData = makeExp(), nExp = 2)
}
\arguments{
\item{expName}{The experiment name, Default: 'exp1'}

\item{expData}{An object of class \code{exp} produced by \code{makeExp}, Default: makeExp()}

\item{nExp}{The number of experiments, Default: 2}
}
\value{
A dataset with experiment run data embedded into the \code{pongaR} library.
}
\description{
A function that runs experiment to produce acoustic emission signals
             due solid particle impacts onto a plate.
}
