% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shearEquation.R
\name{shearEquation}
\alias{shearEquation}
\title{Computing shear equation .}
\usage{
shearEquation(shearCorrectionfactor, poissonRatio)
}
\arguments{
\item{shearCorrectionfactor}{The material Midlin's plate theory shear correction factor.}

\item{poissonRatio}{The material Poisson ratio.}
}
\value{
The value of the shear equation used to calculate the material Midlin's
        plate theory shear correction factor.
}
\description{
A utility function that computes shear equation (R. D. Mindlin, J. Appl. Mech. 18, 38 (1951)).
}
