% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titaniumDD.R
\docType{data}
\name{titaniumDD}
\alias{titaniumDD}
\title{Titanium dispersion data}
\format{Titanium dispersion data as an object of class \code{"list"}.}
\usage{
data(titaniumDD)
}
\description{
Titanium dispersion data calculated by \code{makeDD}.
}
\keyword{datasets}
