% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titaniumEF.R
\docType{data}
\name{titaniumEF}
\alias{titaniumEF}
\title{Titanium excitability function data}
\format{Titanium excitability function data data as an object of class \code{"list"}.}
\usage{
data(titaniumEF)
}
\description{
Titanium excitability function data calculated by \code{makeEF}.
}
\keyword{datasets}
