% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vel2lame.R
\name{vel2lame}
\alias{vel2lame}
\title{Lame parameters from longitudinal and transverse velocities and density.}
\usage{
vel2lame(density = 4460, longVelocity = 6060, tranVelocity = 3230)
}
\arguments{
\item{density}{The material density in kg/m3, Default: 4460}

\item{longVelocity}{The material longitudinal velocity in m/s, Default: 6060}

\item{tranVelocity}{The material transverse velocity in m/s, Default: 3230}
}
\value{
The material first and second Lame parameters in Pa as a \code{list} object.
}
\description{
A utility function that converts longitudinal and transverse velocities and density to Lame parameters.
}
