% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vel2young.R
\name{vel2young}
\alias{vel2young}
\title{Young modulus and Poisson ratio from longitudinal and transverse velocities and density.}
\usage{
vel2young(density = 4460, longVelocity = 6060, tranVelocity = 3230)
}
\arguments{
\item{density}{The material density in kg/m3, Default: 4460}

\item{longVelocity}{The material longitudinal velocity in m/s, Default: 6060}

\item{tranVelocity}{The material transverse velocity in m/s, Default: 3230}
}
\value{
The material Young modulus in Pa and Poisson ratio as a \code{list} object.
}
\description{
A utility function that converts longitudinal and transverse velocities and density
             to Young modulus and Poisson ratio.
}
