% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/young2lame.R
\name{young2lame}
\alias{young2lame}
\title{Lame parameters from Young modulus and Poisson ratio.}
\usage{
young2lame(youngModulus = 121127478933, poissonRatio = 0.3015857)
}
\arguments{
\item{youngModulus}{The material Young modulus in Pa, Default: 121127478933}

\item{poissonRatio}{The material Poisson ratio, Default: 0.3015857}
}
\value{
The material first and second Lame parameters in Pa as a \code{list} object.
}
\description{
A utility function that converts Young modulus and Poisson ratio to Lame parameters.
}
