% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/young2vel.R
\name{young2vel}
\alias{young2vel}
\title{Longitudinal and transverse velocities from density, Young modulus and Poisson ratio.}
\usage{
young2vel(density = 4460, youngModulus = 121127478933,
  poissonRatio = 0.3015857)
}
\arguments{
\item{density}{The material density in kg/m3, Default: 4460}

\item{youngModulus}{The material Young modulus in Pa, Default: 121127478933}

\item{poissonRatio}{The material Poisson ratio, Default: 0.3015857}
}
\value{
The material density in kg/m3, longitudinal and transverse velocities in m/s as
        a \code{list} object.
}
\description{
A utility function that converts density, Young modulus and Poisson ratio
             to longitudinal and transverse velocities.
}
